reorganize the dialog context
