+++ stopcovid/dialog/models/__init__.py
@@ -1,0 +1,1 @@
+SCHEMA_VERSION = 1
