+++ stopcovid/status/reminder_triggers.py
@@ -3,16 +3,7 @@ from dataclasses import dataclass
 from typing import List
 import logging
 
-from sqlalchemy import (
-    Table,
-    MetaData,
-    Column,
-    String,
-    select,
-    exists,
-    and_,
-    insert,
-)
+from sqlalchemy import Table, MetaData, Column, String, select, exists, and_, insert
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.schema import UniqueConstraint
 from sqlalchemy.exc import DatabaseError, IntegrityError
